requirements.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.2
django-filter==1.1.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
phonenumbers==8.11.1
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
